+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# <a href='https://github.com/mrjackwills/staticpi_backend/releases/tag/v0.0.15'>v0.0.15</a>
+### 2024-07-27
+
+### Chores
++ dependencies updated, [f362bdbe](https://github.com/mrjackwills/staticpi_backend/commit/f362bdbe6404cecc26f21816b6481fbf4c967cf0),
++ .devcontainer updated, [929e623c](https://github.com/mrjackwills/staticpi_backend/commit/929e623c530e986a32eb75f0f7e4b222fc5b9d24),
++ Rust 1.80 linting, [698d721e](https://github.com/mrjackwills/staticpi_backend/commit/698d721e8809ad90ab811d566654076863b65d50),
+
+### Fixes
++ Docker files updated, [afc6f8bb](https://github.com/mrjackwills/staticpi_backend/commit/afc6f8bb3a51cbf06f836e6193f1db69de5bd43b),
++ run.sh, [e31e50c4](https://github.com/mrjackwills/staticpi_backend/commit/e31e50c4fdee65f27092454b19fcf8a6e84a681e),
++ connect to postgres with `new_without_pgpass()`, [68c8ca24](https://github.com/mrjackwills/staticpi_backend/commit/68c8ca2460ade42dde99a02611a0c83513f1c87f),
+
 # <a href='https://github.com/mrjackwills/staticpi_backend/releases/tag/v0.0.14'>v0.0.14</a>
 ### 2024-06-17